refactor(client-ui): split utils/LogType.kt -- GameConsoleService gets its own file
The file named LogType.kt actually held three concerns: the LogType
enum (10 lines), the LogEntry data class (5 lines), and the
GameConsoleService object (108 lines of state-holder logic for the
in-app console). File name vs principal content was lying.

Rename the file to GameConsoleService.kt (matches the principal class
the Kotlin convention prefers) and extract LogType + LogEntry into a
fresh LogType.kt that now actually only declares the types the name
implies. No behavioural change; ConsoleWindow / LauncherController /
ProcessLogHandler imports keep resolving through the same package.
